Update and modification
1. Add a mask layer - menu-mask on mobile screen when invoke the menu.
Once click the mask layer, menu will be closed. The purpose is to
prevent unexpected menu and layers behavior on mobile screen.
2. Add CSS rule for the last <a> block of Instagram feed
3. Changed the quality of instagram feed pics to high quality.
4. Modified product thumbnails margin computing formula to line up each
picture on desktop and mobile screen.
5. Modified product main picture presentation method, background-image
to img. The purpose is to allow search engine to capture the pictures
properly.
6. Add method to extend page height to fit the window height when page
height is smaller than window height. Better appearance for short pages.